+++ Middlewares/ptpd-v2.0.0/src/dep/sys_time.c
@@ -38,10 +38,10 @@ void displayStats(const PtpClock *ptpClock)
 	switch (ptpClock->portDS.delayMechanism)
 	{
 		case E2E:
-			printf("path delay: %d nsec\n", ptpClock->currentDS.meanPathDelay.nanoseconds);
+			printf("path delay: %d nsec\n", (int)ptpClock->currentDS.meanPathDelay.nanoseconds);
 			break;
 		case P2P:
-			printf("path delay: %d nsec\n", ptpClock->portDS.peerMeanPathDelay.nanoseconds);
+			printf("path delay: %d nsec\n", (int)ptpClock->portDS.peerMeanPathDelay.nanoseconds);
 			break;
 		default:
 			printf("path delay: unknown\n");
@@ -52,11 +52,11 @@ void displayStats(const PtpClock *ptpClock)
 	/* Offset from master */
 	if (ptpClock->currentDS.offsetFromMaster.seconds)
 	{
-		printf("offset: %d sec\n", ptpClock->currentDS.offsetFromMaster.seconds);
+		printf("offset: %d sec\n", (int)ptpClock->currentDS.offsetFromMaster.seconds);
 	}
 	else
 	{
-		printf("offset: %d nsec\n", ptpClock->currentDS.offsetFromMaster.nanoseconds);
+		printf("offset: %d nsec\n", (int)ptpClock->currentDS.offsetFromMaster.nanoseconds);
 	}
 
 	/* Observed drift from master */
@@ -82,14 +82,14 @@ void setTime(const TimeInternal *time)
 	ts.tv_sec = time->seconds;
 	ts.tv_nsec = time->nanoseconds;
 	ethernetif_ptp_set_time(&ts);
-	DBG("resetting system clock to %d sec %d nsec\n", time->seconds, time->nanoseconds);
+	DBG("resetting system clock to %d sec %d nsec\n", (int)time->seconds, (int)time->nanoseconds);
 }
 
 void updateTime(const TimeInternal *time)
 {
 	struct ptptime_t timeoffset;
 
-	DBGV("updateTime: %d sec %d nsec\n", time->seconds, time->nanoseconds);
+	DBGV("updateTime: %d sec %d nsec\n", (int)time->seconds, (int)time->nanoseconds);
 
 	timeoffset.tv_sec = -time->seconds;
 	timeoffset.tv_nsec = -time->nanoseconds;
@@ -106,7 +106,7 @@ uint32_t getRand(uint32_t randMax)
 
 bool adjFreq(int32_t adj)
 {
-	DBGV("adjFreq %d\n", adj);
+	DBGV("adjFreq %d\n", (int)adj);
 
 	if (adj > ADJ_FREQ_MAX)
 		adj = ADJ_FREQ_MAX;
